requirements.txt
et_xmlfile==2.0.0
OpenOPC-DA @ git+https://github.com/PerEAJ/openopc.git@44cc821bd38a651d82311f4c793c26b752499a77
openpyxl==3.1.5
Pyro5==5.15
pywin32==310
serpent==1.41
